requirements.txt
python-dotenv
streamlit
azure.storage.blob
azure.core
azure-ai-documentintelligence